#178 refactor: extract buttonlist-functions to utils
